The type TomlTableRef is now used everywhere
Functions like "parseStream" used to have value semantics, i.e., they
produced copies. This is no longer the case: they produce GC-managed
objects allocated on the heap. This lead to some changes in the
TomlValue type (specifically, the tableVal type is no longer a
TomlTable, but a TomlTableRef). This allowed to substitute "ptr" to
TomlTable types with "ref", thus removing a lot of calls to "addr".

This change was done mainly to easily implement the function
"getValueFromFullAddr".