Datetime objects are now correctly parsed
This involved a change in the "parseInt" API: now the procedure accepts
an additional parameter of type "LeadingZero", since by default integer
values cannot start with "0" in TOML files. But in a datetime object,
hours, minutes, seconds, months, and days must be 2-wide strings that
are zero-filled. Additional test cases for "parseInt" have been added
accordingly (as well as tests for parseDateTimePart", of course).