Fix validation bugs, bump to v0.4.0, string output
This updates the parser to use the v0.4.0 TOML spec. It also makes sure
that everything runs through the TOML validation suite (but with an
error on double table keys). String output of the TOML data is now also
part of the library, however this is still early in it's implementation
and probably has bugs.