Copied the interface from the JSON module
This redoes the entire interface with TOML objects to be more or less
the exact same as for the JSON module. TOML objects can also be
generated in the same way that JSON is with using `?` and `?*`
operators, along with a fairly rudimentary `parseToml` macro that takes
actual TOML data.